+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add warning to stream functions, documenting problems that can occur when streaming
+  with query-time preloads.
+
 ## v3.0.0
 
 - Test against Elixir 1.18.